move more common functionality into the base plot class so it can be used by bar and box plots (and likely others)
